Create db.models.stacks
